fix: move tile capital-link preview to economy for line budget (#4149)
Extract provinceTileCapitalLinkPreview into colonizethis_economy so
colonizethis_app main stays within the 60_550-line ceiling; update the
build-improvement wide-panel golden for new Tile capital-link rows.

Refs #4149
