+++ app/lib/features/game/flame/overlays/province_detail_overlay_host_support.dart
@@ -3,7 +3,6 @@
 export '../../../../core/services/game_service/try_get_game_map_data.dart'
     show tryGetGameMapData;
 export 'province_detail_overlay_host_support_bonus.dart';
-export 'province_detail_overlay_host_support_tile_capital_link.dart';
 export 'province_detail_overlay_host_support_display.dart';
 export 'province_detail_overlay_host_support_factory.dart';
 export 'province_detail_overlay_host_support_map_data.dart';
