+++ packages/colonizethis_economy/lib/colonizethis_economy.dart
@@ -18,6 +18,7 @@ export 'src/economy/economy_riches_to_treasury.dart';
 export 'src/economy/non_gp_auto_offers.dart';
 export 'src/economy/non_gp_extraction.dart';
 export 'src/economy/province_extraction_snapshot_builder.dart';
+export 'src/economy/province_tile_capital_link_preview.dart';
 export 'src/economy/province_improvable_resource_counts.dart';
 export 'src/economy/projected_cost_engine.dart';
 export 'src/economy/resource_extractor.dart';
